CMSIS: AArch64: gic_v3: add support for multiple thread processors
In the processors, which integrates the multi-threaded cores, the
definition of the affinity fields in MPIDR_EL1 is different, the
Affinity0 is used to indicate Thread ID instead of Core ID.
However, the current implementation hard-code to use Affinity0 to
identify Core ID and then index the Redistributor, so it does't
work anymore on these platforms.

This patch remove the hard-code indenification of Redistributor
and change to use the core/thread's affinity fields to match the
Redistributor.

Signed-off-by: Hou Zhiqiang <Zhiqiang.Hou@nxp.com>