mcux: hal_nxp.cmake: Link drivers with their zephyr components
Modify NXP HAL CMake include_driver_ifdef function to also accept an
optional argument specifying the subsystem to use when linking the
driver. Update usages of include_driver_ifdef to link drivers with their
relevant subystems.

The advantage of this approach is that when code is relocated (IE if a
driver subsystem is relocated to RAM) the SDK driver will also be
relocated if linked with the subsystem itself.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>