[nxp fromlist] boards: rd_rw612_bga: HDLC communication added in the device tree
Add a HDLC communication in the device tree to support an OpenThread
RCP HOST mode in Zephyr for the redfinch component which have two CPUs :
host + radio via an IMU link.
Today, only the OpenThread RCP standalone mode (only one CPU) is
supported in Zephyr.

The west.yml file is updated with the 2 other repositories linkeds with
this commit as follows :
- nxp-zephyr, SHA1 : aa71c6e4373c07d18e827982063c43ac7a3e803c
- hal-nxp, SHA1 : 44724e2efd39a0b85b313f2d2d6849f04831475d

Signed-off-by: Xavier Razavet <xavier.razavet@nxp.com>