Issue: WSW-32956 [WIFI][RW610] Set region fail for RW610
[Description]
When change region and country code using cmd wlan-set-country or wlan-set-regioncode, the actual region and country code in 0x249 cmd is always 0x10 US.
The power value is not changed in this PR, it's still from https://jira.sw.nxp.com/browse/GLC8890-582

[Root Cause]
The the region input in the excel to generate tx power table is always US.

[Fix]
Use corresponding country code and region code to generate corresponding tx power table for 0x249 cmd.
Please be attention, please use the uniform format excel as other chip which contain the region info in all of the sheet and write region info into them,
otherwise the region info in country TLV and power table TLV will be different and the new channel list and new power table will be download fail.

And the power value is not changed in this PR, it's still from:
https://jira.sw.nxp.com/browse/GLC8890-582

How to fill dfs region in the excel?
The value of dfs region can only be 0 to 3.
0 for unset, 1 for FCC, 2 for EU, 3 for JP.
For china CN region, it use the same dfs region as FCC, so it's 1.

[Testing]
The region info can be changed normally after change.

[Potential Side-effect]
NA.

[Porting To]
No.

Signed-off-by: Weijian Shao <weijian.shao@nxp.com>