[ZEP-1906] west.yml: Add hal_st to allowed list
Adding hal_st to the allowed list of modules to pull in from the
upstream.

There are sensor shields based on ST parts enabled in the upstream.
These shields/sensors require the hal_st module. Note that this is
different from the hal_stm32 module which is not included in the
NXP-ZSDK downstream.

Signed-off-by: David Leach <david.leach@nxp.com>