zephyr: rw61x: rework cpu2 monolithic handling
To simplify how coexistence between BLE and 15.4 works, we decided to
choose the cpu2 firmware based on which SoC is used.
If RW612 is used, then we use the combo firmware.
If RW610 is used, then we use the ble only firmware.

MONOLITHIC_BT and MONOLITHIC_IEEE802154 configs have been combined
into a single one: MONOLITHIC_NBR.

We also unify the fw array symbol to `fw_cpu2` and make it transparent
to the connectivity_framework by using the preprocessor.

Signed-off-by: Axel Le Bourhis <axel.lebourhis@nxp.com>