[nxp fromtree][MCUX-83926] codec: port: wm8960: remove faulty volume check for DAC
Fix faulty volume check when setting the DAC volume, as the function
clamps the volume to 100 and then checks if it is superior to
UINT32_MAX/ WM8960_DAC_MAX_VOLUME_vALUE = (0xFFFFFFFF/0xFF) = 16843009,
which is never true and always leads to a kStatus_Fail.
Align with previous checks for other components and set to max value
when mappedVolume would wrap.

Signed-off-by: Aziz Sellami <aziz.sellami@nxp.com>